+++ config/aliases-plate.js
@@ -38,6 +38,7 @@ module.exports = {
   '@udecode/plate-select': 'editor/select',
   '@udecode/plate-selection': 'selection',
   '@udecode/plate-suggestion': 'suggestion',
+  '@udecode/plate-tabbable': 'editor/tabbable',
   '@udecode/plate-table': 'nodes/table',
   '@udecode/plate-test-utils': 'test-utils',
   '@udecode/plate-trailing-block': 'editor/trailing-block',