+++ docs/docs/sandpack/files/reset-node/code-resetBlockTypePlugin.ts
@@ -1,16 +1,26 @@
 export const resetBlockTypePluginCode = `import {
   ELEMENT_BLOCKQUOTE,
+  ELEMENT_CODE_BLOCK,
   ELEMENT_PARAGRAPH,
   ELEMENT_TODO_LI,
   isBlockAboveEmpty,
+  isCodeBlockEmpty,
   isSelectionAtBlockStart,
+  isSelectionAtCodeBlockStart,
   ResetNodePlugin,
+  unwrapCodeBlock,
 } from '@udecode/plate';
 import { MyPlatePlugin } from '../typescript/plateTypes';
 
 const resetBlockTypesCommonRule = {
   types: [ELEMENT_BLOCKQUOTE, ELEMENT_TODO_LI],
   defaultType: ELEMENT_PARAGRAPH,
+};
+
+const resetBlockTypesCodeBlockRule = {
+  types: [ELEMENT_CODE_BLOCK],
+  defaultType: ELEMENT_PARAGRAPH,
+  onReset: unwrapCodeBlock,
 };
 
 export const resetBlockTypePlugin: Partial<MyPlatePlugin<ResetNodePlugin>> = {
@@ -25,6 +35,16 @@ export const resetBlockTypePlugin: Partial<MyPlatePlugin<ResetNodePlugin>> = {
         ...resetBlockTypesCommonRule,
         hotkey: 'Backspace',
         predicate: isSelectionAtBlockStart,
+      },
+      {
+        ...resetBlockTypesCodeBlockRule,
+        hotkey: 'Enter',
+        predicate: isCodeBlockEmpty,
+      },
+      {
+        ...resetBlockTypesCodeBlockRule,
+        hotkey: 'Backspace',
+        predicate: isSelectionAtCodeBlockStart,
       },
     ],
   },