+++ docs/docs/sandpack/plate-versions.ts
@@ -1,8 +1,8 @@
-export const plateVersion = '19.0.7';
+export const plateVersion = '19.4.3';
 export const testUtilsVersion = '13.4.0';
-export const excalidrawVersion = '19.0.3';
-export const juiceVersion = '19.0.3';
-export const selectionVersion = '19.0.5';
-export const cloudVersion = '19.0.3';
-export const cloudUiVersion = '19.0.3';
-export const dndVersion = '19.0.3';
+export const excalidrawVersion = '19.4.2';
+export const juiceVersion = '19.4.2';
+export const selectionVersion = '19.4.2';
+export const cloudVersion = '19.4.2';
+export const cloudUiVersion = '19.4.2';
+export const dndVersion = '19.4.2';