+++ examples/apps/next/src/pages/tabbable.tsx
@@ -1,0 +1,3 @@
+import App from 'examples/src/TabbableApp';
+
+export default App;