+++ packages/nodes/code-block/src/transforms/insertCodeBlock.spec.tsx
@@ -1,0 +1,125 @@
+/** @jsx jsx */
+
+import { PlateEditor } from '@udecode/plate-core';
+import { jsx } from '@udecode/plate-test-utils';
+import { createPlateUIEditor } from '../../../../ui/plate/src/utils/createPlateUIEditor';
+import { createCodeBlockPlugin } from '../createCodeBlockPlugin';
+import { insertCodeBlock } from './insertCodeBlock';
+
+jsx;
+
+describe('insert code block', () => {
+  describe('when selection is at start of block', () => {
+    it('should turn line to code block', () => {
+      const input = ((
+        <editor>
+          <hp>line 1</hp>
+          <hp>
+            <cursor />
+            line 2
+          </hp>
+          <hp>line 3</hp>
+        </editor>
+      ) as any) as PlateEditor;
+
+      const output = ((
+        <editor>
+          <hp>line 1</hp>
+          <hcodeblock>
+            <hcodeline>
+              <cursor />
+              line 2
+            </hcodeline>
+          </hcodeblock>
+          <hp>line 3</hp>
+        </editor>
+      ) as any) as PlateEditor;
+
+      const editor = createPlateUIEditor({
+        editor: input,
+        plugins: [createCodeBlockPlugin()],
+      });
+
+      insertCodeBlock(editor);
+
+      expect(input.children).toEqual(output.children);
+    });
+  });
+
+  describe('when selection is not at start of block', () => {
+    it('should split line at selection and turn latter line to code block', () => {
+      const input = ((
+        <editor>
+          <hp>line 1</hp>
+          <hp>
+            before <cursor />
+            after
+          </hp>
+          <hp>line 3</hp>
+        </editor>
+      ) as any) as PlateEditor;
+
+      const output = ((
+        <editor>
+          <hp>line 1</hp>
+          <hp>before </hp>
+          <hcodeblock>
+            <hcodeline>
+              <cursor />
+              after
+            </hcodeline>
+          </hcodeblock>
+          <hp>line 3</hp>
+        </editor>
+      ) as any) as PlateEditor;
+
+      const editor = createPlateUIEditor({
+        editor: input,
+        plugins: [createCodeBlockPlugin()],
+      });
+
+      insertCodeBlock(editor);
+
+      expect(input.children).toEqual(output.children);
+    });
+  });
+
+  describe('when selection is expanded', () => {
+    it('should do nothing', () => {
+      const input = ((
+        <editor>
+          <hp>line 1</hp>
+          <hp>
+            before <anchor />
+            selection
+            <focus />
+            after
+          </hp>
+          <hp>line 3</hp>
+        </editor>
+      ) as any) as PlateEditor;
+
+      const output = ((
+        <editor>
+          <hp>line 1</hp>
+          <hp>
+            before <anchor />
+            selection
+            <focus />
+            after
+          </hp>
+          <hp>line 3</hp>
+        </editor>
+      ) as any) as PlateEditor;
+
+      const editor = createPlateUIEditor({
+        editor: input,
+        plugins: [createCodeBlockPlugin()],
+      });
+
+      insertCodeBlock(editor);
+
+      expect(input.children).toEqual(output.children);
+    });
+  });
+});