+++ packages/nodes/code-block/src/transforms/insertEmptyCodeBlock.ts
@@ -1,6 +1,5 @@
 import {
   ELEMENT_DEFAULT,
-  getPath,
   getPluginType,
   insertElements,
   isBlockAboveEmpty,
@@ -8,7 +7,7 @@ import {
   PlateEditor,
   Value,
 } from '@udecode/plate-core';
-import { Path } from 'slate';
+import { Path, Range } from 'slate';
 import { CodeBlockInsertOptions } from '../types';
 import { insertCodeBlock } from './insertCodeBlock';
 
@@ -27,7 +26,7 @@ export const insertEmptyCodeBlock = <V extends Value>(
   if (!editor.selection) return;
 
   if (isExpanded(editor.selection) || !isBlockAboveEmpty(editor)) {
-    const selectionPath = getPath(editor, editor.selection);
+    const selectionPath = Range.end(editor.selection).path;
     const insertPath = Path.next(selectionPath.slice(0, level + 1));
     insertElements(
       editor,