+++ packages/nodes/mention/src/types.ts
@@ -1,5 +1,5 @@
 import { Data, NoData } from '@udecode/plate-combobox';
-import { TElement } from '@udecode/plate-core';
+import { PlateEditor, TElement, Value } from '@udecode/plate-core';
 import { CreateMentionNode } from './getMentionOnSelectItem';
 
 export interface TMentionElement extends TElement {
@@ -16,4 +16,7 @@ export interface MentionPlugin<TData extends Data = NoData> {
   insertSpaceAfterMention?: boolean;
   trigger?: string;
   inputCreation?: { key: string; value: string };
+  query?: <V extends Value = Value, E extends PlateEditor<V> = PlateEditor<V>>(
+    editor: E
+  ) => boolean;
 }